Rename Treehouse Compose UI module to include 'host'
Also fold the insets module into this module as an implementation detail.
