--- redwood-treehouse-composeui-insets/src/commonMain/kotlin/app/cash/redwood/treehouse/composeui/insets.kt
@@ -1,23 +1,0 @@
-/*
- * Copyright (C) 2023 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package app.cash.redwood.treehouse.composeui
-
-import androidx.compose.runtime.Composable
-import app.cash.redwood.ui.Margin
-
-/** Return the device's insets in screen density-independent pixels. */
-@Composable
-public expect fun safeAreaInsets(): Margin
