+++ settings.gradle
@@ -75,11 +75,10 @@ include ':redwood-tooling-codegen'
 include ':redwood-tooling-lint'
 include ':redwood-tooling-schema'
 include ':redwood-treehouse'
-include ':redwood-treehouse-composeui'
-include ':redwood-treehouse-composeui-insets'
 include ':redwood-treehouse-guest'
 include ':redwood-treehouse-guest-compose'
 include ':redwood-treehouse-host'
+include ':redwood-treehouse-host-composeui'
 include ':redwood-widget'
 include ':redwood-widget-compose'
 include ':redwood-widget-testing'
